+++ src/components/kepler-gl.js
@@ -245,7 +245,7 @@ function KeplerGlFactory(
         mapboxApiUrl,
         mapState,
         uiState,
-        visState,
+        editor: visState.editor,
         mapStyle,
         mapControls: uiState.mapControls,
         layers,
