+++ src/components/map-container.js
@@ -21,13 +21,11 @@
 // libraries
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
-import styled from 'styled-components';
 import MapboxGLMap from 'react-map-gl';
 import DeckGL from 'deck.gl';
-import WebMercatorViewport from 'viewport-mercator-project';
 
 // components
-import MapPopoverFactory from 'components/map/map-popover';
+import MapTooltipFactory from 'components/map/map-tooltip';
 import MapControlFactory from 'components/map/map-control';
 import {StyledMapContainer} from 'components/common/styled-components';
 
@@ -40,10 +38,6 @@ import {transformRequest} from 'utils/map-style-utils/mapbox-utils';
 
 // default-settings
 import ThreeDBuildingLayer from '../deckgl-layers/3d-building-layer/3d-building-layer';
-
-const Editor = styled(Draw)`
-  display: ${props => props.visible ? 'default' : 'none'};
-`;
 
 const MAP_STYLE = {
   container: {
@@ -60,86 +54,11 @@ const MAPBOXGL_RENDER = 'render';
 const TRANSITION_DURATION = 0;
 
 MapContainerFactory.deps = [
-  MapPopoverFactory,
+  MapTooltipFactory,
   MapControlFactory
 ];
 
-export default function MapContainerFactory(MapPopover, MapControl) {
-  /* eslint-disable complexity */
-  const MapTooltip = React.memo(({
-                             mapState,
-                             hoverInfo,
-                             clicked,
-                             datasets,
-                             interactionConfig,
-                             layers,
-                             mapLayers,
-                             mousePos: {mousePosition, coordinate, pinned}
-                           }) => {
-    if (!mousePosition) {
-      return null;
-    }
-
-    // if clicked something, ignore hover behavior
-    const objectInfo = clicked || hoverInfo;
-    let layerHoverProp = null;
-    let position = {x: mousePosition[0], y: mousePosition[1]};
-
-    if (
-      interactionConfig.tooltip.enabled &&
-      objectInfo &&
-      objectInfo.picked
-    ) {
-      // if anything hovered
-      const {object, layer: overlay} = objectInfo;
-
-      // deckgl layer to kepler-gl layer
-      const layer = layers[overlay.props.idx];
-
-      if (
-        layer.config.isVisible &&
-        layer.getHoverData &&
-        (!mapLayers || mapLayers[layer.id].isVisible)
-      ) {
-
-        // if layer is visible and have hovered data
-        const {config: {dataId}} = layer;
-        const {allData, fields} = datasets[dataId];
-        const data = layer.getHoverData(object, allData);
-        const fieldsToShow = interactionConfig.tooltip.config.fieldsToShow[dataId];
-
-        layerHoverProp = {
-          data,
-          fields,
-          fieldsToShow,
-          layer
-        }
-      }
-    }
-
-    if (pinned || clicked) {
-      // project lnglat to screen so that tooltip follows the object on zoom
-      const viewport = new WebMercatorViewport(mapState);
-      const lngLat = clicked ? clicked.lngLat : pinned.coordinate;
-      position = this._getHoverXY(viewport, lngLat);
-    }
-    return (
-      <div>
-        <MapPopover
-          {...position}
-          layerHoverProp={layerHoverProp}
-          coordinate={interactionConfig.coordinate.enabled && ((pinned || {}).coordinate || coordinate)}
-          freezed={Boolean(clicked || pinned)}
-          onClose={this._onCloseMapPopover}
-          mapW={mapState.width}
-          mapH={mapState.height}
-        />
-      </div>
-    );
-  });
-  MapTooltip.displayName = 'MapTooltip';
-  /* eslint-enable complexity */
-
+export default function MapContainerFactory(MapTooltip, MapControl) {
   class MapContainer extends Component {
     static propTypes = {
       // required
@@ -259,11 +178,6 @@ export default function MapContainerFactory(MapPopover, MapControl) {
     };
 
     /* component render functions */
-    _getHoverXY(viewport, lngLat) {
-      const screenCoord = !viewport || !lngLat ? null : viewport.project(lngLat);
-      return screenCoord && {x: screenCoord[0], y: screenCoord[1]};
-    }
-
     _shouldRenderLayer(layer, data, mapLayers) {
       const isAvailableAndVisible =
         !(mapLayers && mapLayers[layer.id]) || mapLayers[layer.id].isVisible;
@@ -409,7 +323,7 @@ export default function MapContainerFactory(MapPopover, MapControl) {
       const {
         mapState, mapStyle, mapStateActions, mapLayers, layers, MapComponent,
         datasets, mapboxApiAccessToken, mapboxApiUrl, mapControls,
-        uiState, uiStateActions, visState, visStateActions,
+        uiState, uiStateActions, editor, visStateActions,
         hoverInfo, clicked, interactionConfig,
         mousePos
       } = this.props;
@@ -458,6 +372,7 @@ export default function MapContainerFactory(MapPopover, MapControl) {
               mapStyle={mapStyle.bottomMapStyle}
               getCursor={this.props.hoverInfo ? () => 'pointer' : undefined}
               transitionDuration={TRANSITION_DURATION}
+              onMouseMove={this.props.visStateActions.onMouseMove}
             >
               {this._renderOverlay()}
               {this._renderMapboxOverlays()}
@@ -465,10 +380,10 @@ export default function MapContainerFactory(MapPopover, MapControl) {
                 By placing the editor in this map we have to perform fewer checks for css zIndex
                 and fewer updates when we switch from edit to read mode
               */}
-              <Editor
+              <Draw
                 datasets={datasets}
                 editor={uiState.editor}
-                features={visState.editor.features}
+                features={editor.features}
                 isEnabled={isEdit}
                 layers={layers}
                 onDeleteFeature={uiStateActions.deleteFeature}
@@ -497,6 +412,7 @@ export default function MapContainerFactory(MapPopover, MapControl) {
             layers={layers}
             mapLayers={mapLayers}
             mousePos={mousePos}
+            onClose={this._onCloseMapPopover}
           />
         </StyledMapContainer>
       );
