+++ src/reducers/composers.js
@@ -27,9 +27,7 @@ import * as combinedUpdaters from './combined-updaters';
  */
 const actionHandler = {
   [ActionTypes.UPDATE_VIS_DATA]: combinedUpdaters.updateVisDataUpdater,
-  [ActionTypes.ADD_DATA_TO_MAP]: combinedUpdaters.addDataToMapUpdater,
-  [ActionTypes.SET_FEATURES]: combinedUpdaters.setFeaturesUpdater,
-  [ActionTypes.DELETE_FEATURE]: combinedUpdaters.deleteFeatureUpdater
+  [ActionTypes.ADD_DATA_TO_MAP]: combinedUpdaters.addDataToMapUpdater
 };
 
 export default actionHandler;
