+++ src/reducers/vis-state-updaters.js
@@ -1399,6 +1399,12 @@ export function setFeaturesUpdater(state, {features = []}) {
   }
 }
 
+/**
+ * @memberof visStateUpdaters
+ * @param {Object} state `visState`
+ * @param {string} selectedFeatureId feature to delete
+ * @return {Object} nextState
+ */
 export const deleteFeatureUpdater = (state, {payload: selectedFeatureId}) => {
   return selectedFeatureId ? {
     ...state,
