+++ src/styles/base.js
@@ -225,6 +225,10 @@ export const notificationPanelWidth = 240;
 export const notificationPanelItemWidth = notificationPanelWidth - 60;
 export const notificationPanelItemHeight = 60;
 
+// Action Panel
+export const actionPanelWidth = 110;
+export const actionPanelHeight = 32;
+
 export const textTruncate = {
   maxWidth: '100%',
   overflow: 'hidden',
@@ -948,6 +952,10 @@ export const theme = {
   notificationPanelWidth,
   notificationPanelItemWidth,
   notificationPanelItemHeight,
+
+  // Action Panel
+  actionPanelWidth,
+  actionPanelHeight,
 
   // Breakpoints
   breakPoints
